BackendChat/requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
channels==4.1.0
click==8.1.7
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
flake8==7.0.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.0
uritemplate==4.1.1
